message sending on configured timer and test added
